move process cues to process/
