+++ lib/convert-cue-to-dom-tree.js
@@ -1,0 +1,10 @@
+var parseContent = require('./process/parse-content.js');
+
+var convertCueToDOMTree = function(window, cuetext) {
+  if (!window || !cuetext) {
+    return null;
+  }
+  return parseContent(window, cuetext);
+};
+
+module.exports = convertCueToDOMTree;
