+++ lib/vttcue.js
@@ -13,6 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+var convertCueToDOMTree = require('./convert-cue-to-dom-tree.js');
 
 var autoKeyword = "auto";
 var directionSetting = {
@@ -280,7 +281,7 @@ function VTTCue(startTime, endTime, text) {
 
 VTTCue.prototype.getCueAsHTML = function() {
   // Assume WebVTT.convertCueToDOMTree is on the global.
-  return WebVTT.convertCueToDOMTree(window, this.text);
+  return convertCueToDOMTree(window, this.text);
 };
 
 module.exports = VTTCue;
